% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Simulated data from the Subgroup learning in functional regression under the RKHS framework}
\usage{
simdata(n, m, kenel_sigma, oo, gamma)
}
\arguments{
\item{n}{sample size.}

\item{m}{functional individual size.}

\item{kenel_sigma}{the kernel bandwidth of a gaussian kernel.}

\item{oo}{the subset vector}

\item{gamma}{the threshold parameter for default.}
}
\value{
A list with the elements
\item{y}{The response variable.}
\item{x}{The scalar covariate with threshold.}
\item{z}{A vector of covariates.}
\item{s}{A vector of locations or times.}
\item{k}{A matrix of kernel gram.}
}
\description{
The function for simulating data from the Subgroup learning in functional regression under the RKHS framework
}
\examples{

## simulated data
ptm <- proc.time()
n <- 100
m <- 10
kenel_sigma <- 0.2
oo <- c(2, 3)
gamma <- c(-1,1)
dat <- simdata(n,m,kenel_sigma,oo,gamma)
head(dat)
proc.time() - ptm
}
\author{
Xin Guan, Xu Liu and Jinhong You
}
\keyword{simdata}
