% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ls.R
\name{smooth_ls}
\alias{smooth_ls}
\title{Simulated data from the Subgroup learning in functional regression under the RKHS framework}
\usage{
smooth_ls(X, Y, Z, K, lambda, oo, gamma.ini, h)
}
\arguments{
\item{X}{A vector of covariates}

\item{Y}{A vector of response.}

\item{Z}{A scalar covariate with threshold.}

\item{K}{A matrix of kernel gram.}

\item{lambda}{the parameter.}

\item{oo}{the subset vector}

\item{gamma.ini}{the threshold parameter for initial.}

\item{h}{the smooth parameter}
}
\value{
A list with the elements
\item{res}{The residuals.}
\item{b}{The regression patameters}
\item{Yhat}{The estimated response}
\item{etahat}{The estimated eta function.}
\item{epshat}{The estimated eps function.}
}
\description{
The function for simulating data from the Subgroup learning in functional regression under the RKHS framework
}
\examples{

## simulated data
ptm <- proc.time()
n <- 100
m <- 10
kenel_sigma <- 0.2
oo <- c(2, 3)
gamma <- c(-1,1)
data <- simdata(n,m,kenel_sigma,oo,gamma)
X = data$X; Y=data$Y; S=data$S; K=data$K;Z=data$Z
gamma.ini = c(0,0)
lambda = 0.01
h = log(n)*n^(-0.5)
result <- smooth_ls(X, Y, Z, K, lambda, oo, gamma.ini, h)
proc.time() - ptm
}
\author{
Xin Guan, Xu Liu and Jinhong You
}
\keyword{smooth_ls}
